% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_metrics.R
\name{eval.auc}
\alias{eval.auc}
\title{Evaluate Area Under ROC Curve}
\usage{
eval.auc(preds, actual)
}
\arguments{
\item{preds}{vector of predicted probabilities for response (positive class)}

\item{actual}{binary vector of actual response}
}
\value{
numeric calculation of AUC value
}
\description{
Evaluate Area Under ROC Curve
}
\details{
algorithm comes from \url{https://blog.revolutionanalytics.com/2017/03/auc-meets-u-stat.html},
also from \url{https://github.com/mlr-org/mlr/blob/e710dd0ac84b5a6c8197f8902b330d5c5c92b26a/R/measures.R#L778}
}
