% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_metrics.R
\name{eval.roc}
\alias{eval.roc}
\title{Generate Receiver Operating Characteristic Object}
\usage{
eval.roc(preds, actual, plot = TRUE)
}
\arguments{
\item{preds}{vector of predicted probabilities for response (positive class)}

\item{actual}{binary vector of actual response}

\item{plot}{whether to return ROC curve plot; defaults to `TRUE``}
}
\value{
when plot is set to `TRUE`, a ggplot object that visualizes ROC curve
\itemize{
 \item \code{cutoff} will be displayed in color as gradient on the curb
 \item \code{tpr} will be placed on the y-axis representing true positive rate
 \item \code{fpr} will be placed on the x-axis representing false positive rate
}

when plot is set to `FALSE`, a data.frame object
\itemize{
 \item \code{cutoff} different cutoffs utilized to evaluate performance
    (determined by)
 \item \code{tpr} true positive rate \eqn{TP/(TP + FN)}
 \item \code{fpr} false positive rate \eqn{FP/(TN + FP)}
}
}
\description{
Generate Receiver Operating Characteristic Object
}
\details{
object is plottable
}
