% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_metrics.R
\name{eval_metrics}
\alias{eval_metrics}
\alias{eval.auc}
\alias{eval.sens}
\alias{eval.spec}
\title{Binary Classification Model Evaluation Metrics}
\usage{
eval.auc(preds, actual)

eval.sens(preds.class, actual)

eval.spec(preds.class, actual)
}
\arguments{
\item{preds}{vector of predicted probabilities for response (positive class)}

\item{actual}{binary vector of actual response}

\item{preds.class}{binary vector of predicted response values}
}
\value{
numeric calculation of specified metric
}
\description{
access to commonly-used metric calculations
\itemize{
 \item \code{eval.auc} calculates area under ROC curve using algorithm from \url{https://blog.revolutionanalytics.com/2017/03/auc-meets-u-stat.html}
 \item \code{eval.sens} calculates true positive rate \eqn{TPR = TP / (TP + FN)}
 \item \code{eval.spec} calculates true negative rate \eqn{TNR = TN / (TN + FP)}
}
}
