% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_opts.R
\name{get.opt_cutoff}
\alias{get.opt_cutoff}
\title{Find Optimal Cutoff to Utilize in Classification Model}
\usage{
get.opt_cutoff(preds, actual, metric = "roc", custom_func = NULL)
}
\arguments{
\item{preds}{vector of predicted probabilities for response (positive class)}

\item{actual}{binary vector of actual response}

\item{metric}{metric to be optimized for. Defaults to accuracy}

\item{custom_func}{if specified, will use a custom metric to optimize for}
}
\value{
scalar of optimal threshold
}
\description{
Find Optimal Cutoff to Utilize in Classification Model
}
