% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.binary}
\alias{is.binary}
\title{Check if object contains valid response values}
\usage{
is.binary(truth)
}
\arguments{
\item{truth}{object to be tested}
}
\value{
boolean of whether input contains valid response values
}
\description{
Check if object contains valid response values
}
