% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.prob}
\alias{is.prob}
\title{Check if an object contains probabilities}
\usage{
is.prob(pred)
}
\arguments{
\item{pred}{vector of values to be tested}
}
\value{
boolean of whether input contains valid probability values
}
\description{
Check if an object contains probabilities
}
