% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_sampling.R
\name{model_sampling}
\alias{model_sampling}
\alias{sample.strat}
\title{Sampling Methods}
\usage{
sample.strat(response, pct = 0.7)
}
\arguments{
\item{response}{vector of response classifications}

\item{pct}{proportion of response records to include in sample; defaults to 0.7}
}
\value{
vector of indices to pull from response vector
}
\description{
different ways to generate partitions of dataset
\itemize{
 \item \code{sample.strat} ensures sample has adequate representation of response variable
}
}
