% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_formatter.R
\name{str_form}
\alias{str_form}
\title{Format String Values with Substitution.}
\usage{
str_form(txt, ...)
}
\arguments{
\item{txt}{str - text string with substitution keys.}

\item{...}{key/value pairs that should be substituted into `txt`}
}
\value{
`txt` with correctly substituted values
}
\description{
Substitution keys are set with double curly brackets.
Note that all keys detected in `txt` must appear in optional arguments
}
